% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.xts.R
\name{create_test_xts}
\alias{create_test_xts}
\title{Create an xts-object with specified length and interval}
\usage{
create_test_xts(n, from = NULL, interval = "mins", tzone = "GMT",
  randomWalk = TRUE)
}
\arguments{
\item{n}{Required length}

\item{from}{Starting date}

\item{interval}{Required interval}

\item{tzone}{Required interval}

\item{randomWalk}{logical. Should data points generated by using a random walk?}
}
\value{
An xts-object.
}
\description{
Creates an xts-object with specified length and interval
}
\seealso{
\code{\link[xts]{xts}}.
}

