% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc.R
\name{read_asc}
\alias{read_asc}
\title{Read time series data from .asc file format}
\usage{
read_asc(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Provide additional details?}
}
\value{
An xts-object.
}
\description{
Reads the .asc structured time series data output generated by 'W.A.S' and 
creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

