% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc.R
\name{read_asc2}
\alias{read_asc2}
\title{Read time series data from .asc file format}
\usage{
read_asc2(file, attributes = 16, na.strings = "RWLuecke", header = F,
  sep = "", dec = ".", verbose = FALSE, ...)
}
\arguments{
\item{file}{The file to be read.}

\item{attributes}{How many attibutes are provided?}

\item{na.strings}{a character vector of strings which are to be interpreted
as NA values. Blank fields are also considered to be missing values in logical, 
integer, numeric and complex fields.}

\item{header}{a logical value indicating whether the file contains the names
of the variables as its first line. If missing, the value is determined from
the file format: header is set to TRUE if and only if the first row contains 
one fewer field than the number of columns.}

\item{sep}{the field separator character. Values on each line of the file 
are separated by this character.  If sep = "" (the default for read.table) 
the separator is 'white space', that is one or more spaces, tabs, newlines 
or carriage returns.}

\item{dec}{the character used in the file for decimal points.}

\item{verbose}{logical. Provide additional details?}

\item{...}{Further arguments to be passed to read.table.}
}
\value{
An xts-object.
}
\description{
Reads the .asc structured time series data output generated by 'AquaZIS' and 
creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

