% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwlr.R
\name{read_dwlr}
\alias{read_dwlr}
\title{Read time series data from .dwlr file format}
\usage{
read_dwlr(file, verbose = FALSE, sep = ";", dec = ".")
}
\arguments{
\item{file}{The file to be read.}

\item{verbose}{Provide additional details?}

\item{sep}{The field separator character.}

\item{dec}{The character used in the file for decimal points.}
}
\value{
An xts-object.
}
\description{
Reads the .dwlr structured time series data and creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

