% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafana.R
\name{read_grafana}
\alias{read_grafana}
\title{Read time series data from .csv (grafana) file format}
\usage{
read_grafana(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read.}

\item{verbose}{Provide additional details?}
}
\value{
A list of xts-objects.
}
\description{
Reads the .csv structured time series data produced by an grafana export and 
creates a list of xts-objects.
}
\seealso{
\code{\link[xts]{xts}}.
}

