% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosim.R
\name{read_kosim}
\alias{read_kosim}
\title{Read time series data from .klzc file format}
\usage{
read_kosim(file)
}
\arguments{
\item{file}{The file to be read}
}
\value{
An xts-object.
}
\description{
Reads the .klzc structured time series data output generated by 'KOSIM' and 
creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

