% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab.R
\name{read_mat}
\alias{read_mat}
\title{Read time series data from .mat file format}
\usage{
read_mat(file, data_cols = 2, time_col = 1)
}
\arguments{
\item{file}{The file to be read.}

\item{data_cols}{Specifies the columns the data is stored.}

\item{time_col}{Specifies the column the time is stored.}
}
\value{
An xts-object.
}
\description{
Reads the .mat structured time series data output generated by 'Matlab' and
creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

