% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{read_md}
\alias{read_md}
\title{Read rainfall time series data of type .md}
\usage{
read_md(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read.}

\item{verbose}{logical. Provide additional details?}
}
\value{
An xts-object.
}
\description{
Reads the .md structured rainfall time series data output usually used by 
German stormwater quality models and creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

