% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{read_moment}
\alias{read_moment}
\title{Read data from MOMENT}
\usage{
read_moment(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Should informative outputs printed during function evaluation?}
}
\value{
A list of xts-objects or data.frames.
}
\description{
Reads the .csv (time series) or .mdb (results) structured data output generated by the stormwater 
quality model 'MOMENT' and creates a list of xts-objects or a data.frames, respectively.
}
\seealso{
\code{\link{xts}}.
}

