% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivus.R
\name{read_nivus}
\alias{read_nivus}
\title{Read time series data from NIVUS}
\usage{
read_nivus(file, ncol = 4, parameter = NULL, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read.}

\item{ncol}{The number of columns (i.e., parameter) to be read.}

\item{parameter}{Vector of names of parameters to be read.}

\item{verbose}{logical. Provide additional details?}
}
\value{
A list of xts-objects.
}
\description{
Reads NIVUS's structured .txt or .gz time series data and creates a list of
xts-objects.
}
\seealso{
\code{\link[xts]{xts}}.
}

