% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott.R
\name{read_ott}
\alias{read_ott}
\title{Read time series data from OTT}
\usage{
read_ott(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Should informative outputs printed during function evaluation?}
}
\value{
A list of xts-objects.
}
\description{
Reads Ott's structured .OML or .gz time series data and creates a list of 
xts-objects.
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[XML]{xmlInternalTreeParse}}.
}

