% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{read_reg}
\alias{read_reg}
\title{Read rainfall time series data from .reg file format}
\usage{
read_reg(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Provide additional details?}
}
\value{
A list of xts-objects.
}
\description{
Reads the .reg structured rainfall time series data output usually used by 
the stormwater quality model 'MOMENT' and creates a xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

