% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm.R
\name{read_swmm_dat}
\alias{read_swmm_dat}
\title{Read time series data from SWMM}
\usage{
read_swmm_dat(file, header = FALSE, sep = "", dec = ".", skip = 1,
  stringsAsFactors = F, fileEncoding = "UTF-8",
  format = "\%m/\%d/\%Y \%H:\%M", tz = "GMT", verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{header}{s. read.table}

\item{sep}{s. read.table}

\item{dec}{s. read.table}

\item{skip}{s. read.table}

\item{stringsAsFactors}{s. read.table}

\item{fileEncoding}{s. read.table}

\item{format}{s. read.table}

\item{tz}{Sets the time zone.}

\item{verbose}{logical. Provide additional details?}
}
\value{
An xts-object.
}
\description{
Reads the .csv structured time series data output generated by the stormwater 
management model 'SWMM' and creates an xts-object. For reading the binary output 
please check the swmmr-package (https://github.com/dleutnant/swmmr).
}
\seealso{
\code{\link[xts]{xts}}.
}

