% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsf.R
\name{read_tsf}
\alias{read_tsf}
\title{Read time series data from .tsf file format}
\usage{
read_tsf(file)
}
\arguments{
\item{file}{The file to be read}
}
\value{
An xts-object.
}
\description{
Reads the .tsf structured time series data (e.g., PCSWMM) and creates an xts-object.
}
\seealso{
\code{\link[xts]{xts}}.
}

