% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvf.R
\name{read_uvf}
\alias{read_uvf}
\title{Read time series data from .uvf file format}
\usage{
read_uvf(file, verbose = TRUE, return.comments = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Provide additional details?}

\item{return.comments}{logical. Should comments within the file be returned?}
}
\value{
A list of xts-objects.
}
\description{
Reads the .uvf structured time series data and creates a list of xts-objects.
}
\seealso{
\code{\link[xts]{xts}}.
}

