% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtw.R
\name{read_wtw}
\alias{read_wtw}
\title{Read time series data from WTW}
\usage{
read_wtw(file, verbose = FALSE)
}
\arguments{
\item{file}{The file to be read}

\item{verbose}{logical. Provide additional details?}
}
\value{
A list of xts-objects.
}
\description{
Reads WTW's structured .csv time series data and creates a list of xts-objects.
}
\seealso{
\code{\link[xts]{xts}}.
}

