% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{write_md}
\alias{write_md}
\title{Write rainfall time series to .md file format.}
\usage{
write_md(xts, file, verbose = FALSE)
}
\arguments{
\item{xts}{The xts object to be written}

\item{file}{A character string naming the file to write to.}

\item{verbose}{logical. Provide additional details?}
}
\description{
Writes the .md structured rainfall time series data usually used by 
German stormwater quality models.
}
\seealso{
\code{\link[xts]{xts}}.
}

