% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm.R
\name{write_swmm_dat}
\alias{write_swmm_dat}
\title{Write xts-objects to .dat file format}
\usage{
write_swmm_dat(xts, file, is_rainfall = FALSE, seconds = -60,
  is_caldata = FALSE, digits = 6)
}
\arguments{
\item{xts}{The xts object to be written.}

\item{file}{A character string naming the file to write to.}

\item{is_rainfall}{logical. TRUE shifts coredata values with specified 
seconds (default is -60)}

\item{seconds}{The seconds to shift to.}

\item{is_caldata}{logical. Modifies the header if .dat output.}

\item{digits}{Number of digits to write.}
}
\description{
Writes a xts-object to .dat structured time series data format usually 
interpreted by the stormwater management model 'SWMM'.
}
\seealso{
\code{\link{xts}}, \code{\link{write.table}}.
}

