% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsf.R
\name{write_tsf}
\alias{write_tsf}
\title{Write xts-objects to .tsf file format}
\usage{
write_tsf(xts, file, ID, parameter, unit)
}
\arguments{
\item{xts}{The xts-object object to be written.}

\item{file}{A character string naming the file to write to.}

\item{ID}{Sets IDs of time series data.}

\item{parameter}{Sets parameters of the time series data.}

\item{unit}{Sets units of the time series data.}
}
\description{
Writes xts objects to .tsf structured time series data.
}
\seealso{
\code{\link{xts}}.
}

