% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvf.R
\name{write_uvf}
\alias{write_uvf}
\title{Write xts- or matrix-objects to .uvf file format}
\usage{
write_uvf(data, file, function.interpretation, parameter, id, measuring.unit,
  arguments.unit = "time", easting = NA, northing = NA, elevation = NA,
  verbose = TRUE)
}
\arguments{
\item{data}{The data object to be written}

\item{file}{A character string naming the file to write to.}

\item{function.interpretation}{The type of function interpretation. Must be: 'Linie', 'Summenlinie', 'Blockanfang', 'Blockende' or 'Punkt'}

\item{parameter}{Name of parameter}

\item{id}{Number of identification}

\item{measuring.unit}{Unit of measurement}

\item{arguments.unit}{Unit of argument}

\item{easting}{Easting}

\item{northing}{Northing}

\item{elevation}{Elevation}

\item{verbose}{logical. Provide additional details?}
}
\description{
Writes xts or matrix objects to .uvf structured time series data.
}
\seealso{
\code{\link{xts}}.
}

