% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xts.R
\name{write_xts}
\alias{write_xts}
\title{Write xts objects to file}
\usage{
write_xts(xts, file = "", format = "\%Y-\%m-\%d \%H:\%M:\%S", sep = ";",
  dec = ".")
}
\arguments{
\item{xts}{The xts object to write.}

\item{file}{A connection, or a character string naming the file to write to.
If file is "", print to the standard output connection.}

\item{format}{The time format.}

\item{sep}{The field separator string. Values within each row of x are 
separated by this string.}

\item{dec}{The string to use for decimal points in numeric or complex 
columns: must be a single character.}
}
\description{
Writes xts objects to file
}
\seealso{
\code{\link[xts]{xts}}.
}

